requirements.txt
bs4
lxml
nltk
tensorflow-gpu==1.13.1
spacy
gensim
javalang
nltk
numpy
scrapy
textdistance[extra]
pathos